requirements.txt
decorator==4.0.2
dict.sorted==1.0.0
Django==1.8.3
django-appconf==1.0.1
django-compressor==1.5
django-extensions==1.5.5
djangorestframework==3.2.2
drf-nested-routers==0.9.0
drf-toolbox==0.1.7
ipython==4.0.0
ipython-genutils==0.1.0
path.py==7.6
pexpect==3.3
pickleshare==0.5
Pillow==2.9.0
pytz==2015.4
simplegeneric==0.8.1
six==1.9.0
traitlets==4.0.0
Werkzeug==0.10.4
wheel==0.24.0
